atf-sh: Avoid an unnecessary fork() on startup
`expr ${OPTIND} - 1` can be replaced with $((OPTIND - 1)) to avoid spawning
a new processes. While it's unlikely to make much of a performance
difference for a single execution, it happens once for ever test case so
it will add up for a full testsuite run.

Submitted upstream as https://github.com/freebsd/atf/pull/6