vm_fault_quick_hold_pages: check cap permissions
Add a vm_cap_allows_prot() to check if a capability has permissions
matching a given vm_prot_t and use it in vm_fault_quick_hold_pages.

Also use vm_cap_allows_prot in useracc and as a result drop the
requirement that the capability be global.  There is reason to belive
that most capabilities will be local in a co-process world.

Fixes #735