purecap-kernel: Refactor vmem to enable capability handling per-arena.
When creating a new arena, the flag VMEM_MEMORY_ARENA can be passed.
This will enable capability bounds enforcement, representability
size and alignent adjustments and checks.
The signature of vmem_create and vmem_init are updated to accept arena flags,
and hidden behind macros to limit the diff to upstream.